Use data in moduels, instead of params.pp
