+++ hiera.yaml
@@ -1,0 +1,16 @@
+---
+version: 5
+
+defaults:
+  datadir: 'data'
+  data_hash: 'yaml_data'
+
+hierarchy:
+  - name: 'Major Version'
+    path: "os/%{facts.os.name}-%{facts.os.release.major}.yaml"
+  - name: 'Distribution Name'
+    path: "os/%{facts.os.name}.yaml"
+  - name: 'Operating System Family'
+    path: "family/%{facts.os.family}.yaml"
+  - name: 'Common'
+    path: 'common.yaml'
