+++ manifests/package.pp
@@ -3,18 +3,17 @@
 #
 # Install required packages
 #
-class caddy::package inherits caddy {
-
+class caddy::package {
   Exec {
     path    => '/usr/bin:/usr/sbin:/bin:/usr/local/bin',
   }
 
   $caddy_url        = 'https://caddyserver.com/download/linux'
-  $caddy_dl_url     = "${caddy_url}/${caddy::params::arch}?plugins=${caddy::caddy_features}&license=${caddy::caddy_license}&telemetry=${caddy::caddy_telemetry}"
-  $caddy_dl_dir     = "${caddy::params::caddy_tmp_dir}/caddy_linux_${$caddy::params::arch}_custom.tar.gz"
-  $caddy_dl_command = $caddy::params::caddy_license ? {
+  $caddy_dl_url     = "${caddy_url}/${caddy::arch}?plugins=${caddy::caddy_features}&license=${caddy::caddy_license}&telemetry=${caddy::caddy_telemetry}"
+  $caddy_dl_dir     = "${caddy::caddy_tmp_dir}/caddy_linux_${$caddy::arch}_custom.tar.gz"
+  $caddy_dl_command = $caddy::caddy_license ? {
     'personal'   => "curl -o ${caddy_dl_dir} '${caddy_dl_url}'",
-    'commercial' => "curl -o ${caddy_dl_dir} '${caddy_dl_url}' --user ${caddy::params::caddy_account_id}:${caddy::params::caddy_api_key}"
+    'commercial' => "curl -o ${caddy_dl_dir} '${caddy_dl_url}' --user ${caddy::caddy_account_id}:${caddy::caddy_api_key}"
   }
 
   exec { 'install caddy':
@@ -23,7 +22,7 @@ class caddy::package inherits caddy {
   }
 
   exec { 'extract caddy':
-    command => "tar -zxf ${caddy::params::caddy_tmp_dir}/caddy_linux_${$caddy::params::arch}_custom.tar.gz -C ${caddy::install_path} 'caddy'",
+    command => "tar -zxf ${caddy::caddy_tmp_dir}/caddy_linux_${$caddy::arch}_custom.tar.gz -C ${caddy::install_path} 'caddy'",
     creates => "${caddy::install_path}/caddy",
     require => Exec['install caddy'],
   }
@@ -41,5 +40,4 @@ class caddy::package inherits caddy {
     require     => File["${caddy::install_path}/caddy"],
     refreshonly => true,
   }
-
 }
